# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puraani Jeans APIs",
        "description": "Puraani Jeans API documentation",
        "contact": {
            "email": "developer@puranijeans.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": ""
    },
    "servers": [
        {
            "url": "http://safeblockcom.test/",
            "description": "local"
        }
    ],
    "paths": {
...
        "/v1/send/otp": {
            "post": {
                "tags": [
                    "Auth Verification"
                ],
                "summary": "Send Otp",
                "operationId": "sendOtp",
                "requestBody": {
                    "description": "Send OTP for provided phone number",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "country_code": {
                                        "description": "Country code ie. 92 for Pakistan",
                                        "type": "number",
                                        "example": 92
                                    },
                                    "phone_number": {
                                        "description": "Phone Number",
                                        "type": "string",
                                        "example": 3002927320
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": []
                    }
                ]
            }
        },
        "/v1/verify/otp": {
            "post": {
                "tags": [
                    "Auth Verification"
                ],
                "summary": "Verify Otp",
                "operationId": "verifyOtp",
                "requestBody": {
                    "description": "Verify OTP",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "otp": {
                                        "description": "6 Ditig OTP Sent to your mobile number",
                                        "type": "string",
                                        "example": "000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": []
                    }
                ]
            }
        },
        "/v1/resend/otp": {
            "post": {
                "tags": [
                    "Auth Verification"
                ],
                "summary": "Resend Otp",
                "operationId": "resendOtp",
                "requestBody": {
                    "description": "Resend OTP",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "network_id": {
                                        "description": "Network Id of customer's phone number",
                                        "type": "number",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": []
                    }
                ]
            }
        },
...
        "/v1/product/{productId}": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Product detail",
                "operationId": "getProductDetail",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Get product detail",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "referrer_type": {
                                        "description": "referrer type",
                                        "type": "string",
                                        "example": "Featured Product"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "user_access_token": {
                "type": "apiKey",
                "description": "user access token",
                "name": "Authorization",
                "in": "header"
            },
            "client_id": {
                "type": "apiKey",
                "description": "client id",
                "name": "x-client-id",
                "in": "header"
            },
            "client_secret": {
                "type": "apiKey",
                "description": "client secret",
                "name": "x-client-token",
                "in": "header"
            }
        }
    },
...
}